% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AreCohortEventsRestrictedByVisit.R
\name{areCohortEventsRestrictedByVisit}
\alias{areCohortEventsRestrictedByVisit}
\title{Check if Cohort Events are Restricted by Visit}
\usage{
areCohortEventsRestrictedByVisit(cohortDefinition)
}
\arguments{
\item{cohortDefinition}{A list representing the OHDSI Circe cohort definition.}
}
\value{
A logical value indicating whether cohort events are restricted by visit occurrence.
}
\description{
This function checks if cohort events are restricted by visit occurrence in an OHDSI Circe cohort definition.
}
