% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckIfObjectExistsInNestedList.R
\name{checkIfObjectExistsInNestedList}
\alias{checkIfObjectExistsInNestedList}
\title{Check if an Object Exists in a Nested List}
\usage{
checkIfObjectExistsInNestedList(nestedList, object)
}
\arguments{
\item{nestedList}{A nested list in which to search for the object.}

\item{object}{A character string specifying the name of the object to search for.}
}
\value{
A logical value indicating whether the object exists within the nested list.
}
\description{
This function checks whether a specified object exists within a nested list.
}
