% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CirceComparator-package.R, R/CirceComparator.R
\docType{package}
\name{CirceComparator-package}
\alias{CirceComparator-package}
\title{CirceComparator: CirceComparator: To enable comparison of two or more circe compatible cohort definitions.}
\description{
The CirceComparator R package performs pair wise comparison of one or more cohort definitions generated with circe to help surface similarities and differences.

The CirceComparator R package performs pair wise comparison of one or more cohort definitions generated with circe to help surface similarities and differences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CirceComparator/}
  \item \url{https://github.com/OHDSI/CirceComparator}
  \item Report bugs at \url{https://github.com/OHDSI/CirceComparator/issues}
}


Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CirceComparator/}
  \item \url{https://github.com/OHDSI/CirceComparator}
  \item Report bugs at \url{https://github.com/OHDSI/CirceComparator/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
