% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertConceptSetExpressionToDataFrame.R
\name{convertConceptSetExpressionToDataFrame}
\alias{convertConceptSetExpressionToDataFrame}
\title{convert a concept set expression R list object into a data frame object}
\usage{
convertConceptSetExpressionToDataFrame(conceptSetExpression)
}
\arguments{
\item{conceptSetExpression}{An R object that is concept set expression generated by Circe.}
}
\value{
Returns a tibble data frame.
}
\description{
convert a concept set expression R list object into a data frame object
}
