% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractConceptSetsInCohortDefinition.R
\name{extractConceptSetsInCohortDefinition}
\alias{extractConceptSetsInCohortDefinition}
\title{Get Concept Set Definition Set from Cohort Definition}
\usage{
extractConceptSetsInCohortDefinition(cohortDefinition)
}
\arguments{
\item{cohortDefinition}{A list representing the OHDSI Circe cohort definition.}
}
\value{
A data frame
}
\description{
This function returns a data frame with conceptSetId, conceptSetName, etc
}
