% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/GetContinuousPriorObservationPeriodRequirement.R
\name{getContinuousPriorObservationPeriodRequirement}
\alias{getContinuousPriorObservationPeriodRequirement}
\title{Get Continuous Prior Observation Period Requirement}
\usage{
getContinuousPriorObservationPeriodRequirement(cohortDefinition)
}
\arguments{
\item{cohortDefinition}{A list representing the OHDSI Circe cohort definition.}
}
\value{
A list containing the prior and post days of the observation window.
}
\description{
This function retrieves the continuous prior observation period requirements from an OHDSI Circe cohort definition.
}
