% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDomainsInEntryEvents.R
\name{getDomainsInEntryEvents}
\alias{getDomainsInEntryEvents}
\title{Get Domains in Cohort Entry Events}
\usage{
getDomainsInEntryEvents(cohortDefinition)
}
\arguments{
\item{cohortDefinition}{A list representing the OHDSI Circe cohort definition.}
}
\value{
A list containing the unique domains, the number of unique domains, and a tibble indicating the presence of each domain.
}
\description{
This function retrieves the unique domains present in the cohort entry events of an OHDSI Circe cohort definition.
}
