% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInclusionRuleQualifyingEventLimit.R
\name{getInclusionRuleQualifyingEventLimit}
\alias{getInclusionRuleQualifyingEventLimit}
\title{Get Inclusion Rule Qualifying Event Limit}
\usage{
getInclusionRuleQualifyingEventLimit(cohortDefinition)
}
\arguments{
\item{cohortDefinition}{A list representing the OHDSI Circe cohort definition.}
}
\value{
A character string representing the inclusion rule qualifying event limit.
}
\description{
This function retrieves the inclusion rule qualifying event limit from an OHDSI Circe cohort definition.
}
