% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIndexConceptSetsInEntryEvents.R
\name{getIndexConceptSetsInEntryEvents}
\alias{getIndexConceptSetsInEntryEvents}
\title{Get Concept Set used in Cohort Entry Events}
\usage{
getIndexConceptSetsInEntryEvents(cohortDefinition)
}
\arguments{
\item{cohortDefinition}{A list representing the OHDSI Circe cohort definition.}
}
\value{
A list containing the unique domains, the number of unique domains, and a tibble indicating the presence of each domain.
}
\description{
This function retrieves the concept set id for the concept sets used in entry event criteria.
}
