% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInitialEventLimit.R
\name{getInitialEventLimit}
\alias{getInitialEventLimit}
\title{Get Initial Event Limit}
\usage{
getInitialEventLimit(cohortDefinition)
}
\arguments{
\item{cohortDefinition}{A list representing the OHDSI Circe cohort definition.}
}
\value{
A character string representing the initial event limit used in the primary criteria.
}
\description{
This function retrieves the initial event limit from the primary criteria of an OHDSI Circe cohort definition.
}
