% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/GetInitialEventRestrictionAdditionalCriteriaLimit.R
\name{getInitialEventRestrictionAdditionalCriteriaLimit}
\alias{getInitialEventRestrictionAdditionalCriteriaLimit}
\title{Get Initial Event Restriction Additional Criteria Limit}
\usage{
getInitialEventRestrictionAdditionalCriteriaLimit(cohortDefinition)
}
\arguments{
\item{cohortDefinition}{A list representing the OHDSI Circe cohort definition.}
}
\value{
A character string representing the limit value for additional criteria in initial event restriction.
If no additional criteria exist, the default value is "All".
}
\description{
This function retrieves the limit value for additional criteria used in initial event restriction of an OHDSI Circe cohort definition.
}
