% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNumberOfCohortEntryEvents.R
\name{getNumberOfCohortEntryEvents}
\alias{getNumberOfCohortEntryEvents}
\title{Get Number of Cohort Entry Events}
\usage{
getNumberOfCohortEntryEvents(cohortDefinition)
}
\arguments{
\item{cohortDefinition}{A list representing the OHDSI Circe cohort definition.}
}
\value{
An integer representing the number of cohort entry events.
}
\description{
This function retrieves the number of entry events defined in the primary criteria of an OHDSI Circe cohort definition.
}
