% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNumberOfConceptSets.R
\name{getNumberOfConceptSets}
\alias{getNumberOfConceptSets}
\title{Get Number of Concept Sets}
\usage{
getNumberOfConceptSets(cohortDefinition)
}
\arguments{
\item{cohortDefinition}{A list representing the OHDSI Circe cohort definition.}
}
\value{
An integer representing the number of concept sets.
}
\description{
This function retrieves the number of concept sets defined in an OHDSI Circe cohort definition.
}
