% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNumberOfInclusionRules.R
\name{getNumberOfInclusionRules}
\alias{getNumberOfInclusionRules}
\title{Get Number of Inclusion Rules}
\usage{
getNumberOfInclusionRules(cohortDefinition)
}
\arguments{
\item{cohortDefinition}{A list representing the OHDSI Circe cohort definition.}
}
\value{
An integer representing the number of inclusion rules.
}
\description{
This function retrieves the number of inclusion rules defined in an OHDSI Circe cohort definition.
}
