% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetWhereAnObjectExistsInNestedList.R
\name{getWhereAnObjectExistsInNestedList}
\alias{getWhereAnObjectExistsInNestedList}
\title{Identify Locations of an Object in a Nested List}
\usage{
getWhereAnObjectExistsInNestedList(nestedList, object)
}
\arguments{
\item{nestedList}{A nested list in which to search for the object.}

\item{object}{A character string specifying the name of the object to search for.}
}
\value{
A tibble indicating the locations where the object exists within the nested list.
}
\description{
This function identifies the locations of a specified object within a nested list.
}
