% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HasInitialEventRestrictionAdditionalCriteria.R
\name{hasInitialEventRestrictionAdditionalCriteria}
\alias{hasInitialEventRestrictionAdditionalCriteria}
\title{Check for Initial Event Restriction Additional Criteria}
\usage{
hasInitialEventRestrictionAdditionalCriteria(cohortDefinition)
}
\arguments{
\item{cohortDefinition}{A list representing the OHDSI Circe cohort definition.}
}
\value{
A logical value indicating whether additional criteria for initial event restrictions exist.
}
\description{
This function checks if there are additional criteria for initial event restrictions in an OHDSI Circe cohort definition.
}
