% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseCohortDefinitionSpecifications.R
\name{parseCohortDefinitionSpecifications}
\alias{parseCohortDefinitionSpecifications}
\title{Parse OHDSI Circe Cohort Definition Specifications}
\usage{
parseCohortDefinitionSpecifications(cohortDefinition)
}
\arguments{
\item{cohortDefinition}{A list representing the OHDSI Circe cohort definition.}
}
\value{
A tibble containing the parsed cohort definition specifications.
}
\description{
This function parses various specifications from an OHDSI Circe cohort definition, including censor window,
collapse settings, cohort exit strategy, and several other cohort entry event criteria.
}
