% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadCensorWindow.R
\name{readCensorWindow}
\alias{readCensorWindow}
\title{Parse Censor Window from OHDSI Circe Cohort Definition}
\usage{
readCensorWindow(cohortDefinition)
}
\arguments{
\item{cohortDefinition}{A list representing the OHDSI Circe cohort definition.}
}
\value{
A list containing the parsed censor window start and end dates.
}
\description{
This function parses the censor window dates from an OHDSI Circe cohort definition.
}
