% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadCohortExit.R
\name{readCohortExit}
\alias{readCohortExit}
\title{Parse Cohort Exit Strategy from OHDSI Circe Cohort Definition}
\usage{
readCohortExit(cohortDefinition)
}
\arguments{
\item{cohortDefinition}{A list representing the OHDSI Circe cohort definition.}
}
\value{
A list containing the parsed cohort exit strategy, including \code{exitStrategy},
\code{dateOffSetField}, \code{dateOffSet}, \code{drugCodeSetId}, \code{persistenceWindow}, and \code{surveillanceWindow}.
}
\description{
This function parses the cohort exit strategy from an OHDSI Circe cohort definition.
}
