% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadCollapseSettings.R
\name{readCollapseSettings}
\alias{readCollapseSettings}
\title{Parse Collapse Settings from OHDSI Circe Cohort Definition}
\usage{
readCollapseSettings(cohortDefinition)
}
\arguments{
\item{cohortDefinition}{A list representing the OHDSI Circe cohort definition.}
}
\value{
A list containing the parsed collapse settings, including \code{collapseType} and \code{eraPad}.
}
\description{
This function parses the collapse settings from an OHDSI Circe cohort definition.
}
