% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringPresentInCohortDefinitionText.R
\name{stringPresentInCohortDefinitionText}
\alias{stringPresentInCohortDefinitionText}
\title{Check if a string is present in a cohort definition text}
\usage{
stringPresentInCohortDefinitionText(cohortDefinition, textToSearch)
}
\arguments{
\item{cohortDefinition}{A list representing the cohort definition, which will be converted to JSON.}

\item{textToSearch}{A character string to search for within the cohort definition text.}
}
\value{
A logical value indicating whether the textToSearch is present in the cohort definition text.
}
\description{
This function checks if a given string is present within the text of a cohort definition.
}
