% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CirceComparator-package.R, R/CirceComparator.R
\docType{package}
\name{CirceComparator-package}
\alias{CirceComparator-package}
\title{CirceComparator: CirceComparator: Review or Compare Circe Cohort Expressions}
\description{
The CirceComparator R package provides detailed insights into a single Circe cohort expression, including concept sets, entry event criteria, and exit criteria, all presented in tabular form. It also supports pairwise comparison of multiple Circe cohort definitions to highlight similarities and differences.

The CirceComparator R package provides detailed insights into a single Circe cohort expression, including concept sets, entry event criteria, and exit criteria, all presented in tabular form. It also supports pairwise comparison of multiple Circe cohort definitions to highlight similarities and differences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CirceComparator/}
  \item \url{https://github.com/OHDSI/CirceComparator}
  \item Report bugs at \url{https://github.com/OHDSI/CirceComparator/issues}
}


Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CirceComparator/}
  \item \url{https://github.com/OHDSI/CirceComparator}
  \item Report bugs at \url{https://github.com/OHDSI/CirceComparator/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
